refactor: drop support for action parameters ba...
...sed connector generator and move utility classes

This removes `SwaggerActionParamsConnectorGenerator` that's no longer
being used and moves the *Helper classes to `...swagger.util` package
removing any unused methods.
