+++ app/rest/connector-generator/src/main/java/io/syndesis/connector/generator/swagger/util/InMemoryUrlStreamHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.connector.generator.swagger;
+package io.syndesis.connector.generator.swagger.util;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
