+++ app/rest/connector-generator/src/test/java/io/syndesis/connector/generator/swagger/BaseSwaggerGeneratorExampleTest.java
@@ -52,8 +52,6 @@ abstract class BaseSwaggerGeneratorExampleTest extends AbstractSwaggerConnectorT
 
         final Connector generated = generator().generate(SWAGGER_TEMPLATE, connectorSettings);
 
-        Json.mapper().writeValue(System.out, generated);
-
         final Map<String, String> generatedConfiguredProperties = generated.getConfiguredProperties();
         final String generatedSpecification = generatedConfiguredProperties.get("specification");
