--- app/rest/connector-generator/src/test/java/io/syndesis/connector/generator/swagger/SwaggerActionParamsGeneratorExampleTests.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2016 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.syndesis.connector.generator.swagger;
-
-import java.io.IOException;
-import java.util.Arrays;
-
-import io.syndesis.connector.generator.ConnectorGenerator;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
-
-@RunWith(Parameterized.class)
-public class SwaggerActionParamsGeneratorExampleTests extends BaseSwaggerGeneratorExampleTest {
-
-    public SwaggerActionParamsGeneratorExampleTests(final String name) throws IOException {
-        super("param", name);
-    }
-
-    @Test
-    @Override
-    public void shouldGenerateAsExpected() throws IOException {
-        super.shouldGenerateAsExpected();
-    }
-
-    @Override
-    /* default */ ConnectorGenerator generator() {
-        return new SwaggerActionParamsConnectorGenerator();
-    }
-
-    @Parameters(name = "{0}")
-    public static Iterable<String> parameters() {
-        return Arrays.asList("concur", "petstore", "reverb", "basic_auth");
-    }
-
-}
