--- app/rest/connector-generator/src/test/resources/swagger/basic_auth.param_connector.json
@@ -1,133 +1,0 @@
-{
-  "name": "HTTP Basic Authentication Sample",
-  "description": "Sample Swagger file with HTTP Basic Authentication",
-  "connectorGroup": {
-    "id": "swagger-connector-template"
-  },
-  "connectorGroupId": "swagger-connector-template",
-  "configuredProperties": {
-    "specification": "{\"basePath\" : \"/api\", \"host\" : \"sample.io\", \"info\" : {\"description\" : \"Sample Swagger file with HTTP Basic Authentication\", \"license\" : {\"name\" : \"Apache 2.0\", \"url\" : \"http://www.apache.org/licenses/LICENSE-2.0.html\"}, \"title\" : \"HTTP Basic Authentication Sample\", \"version\" : \"1.0.0\"}, \"paths\" : {\"/operation\" : {\"get\" : {\"operationId\" : \"operation-0\", \"parameters\" : [ ], \"produces\" : [ \"application/json\" ], \"responses\" : {\"200\" : {\"description\" : \"All OK\"}}}}}, \"schemes\" : [ \"http\" ], \"securityDefinitions\" : {\"username_password\" : {\"type\" : \"basic\"}}, \"swagger\" : \"2.0\"}",
-    "authenticationType": "basic",
-    "basePath": "/api",
-    "host": "http://sample.io"
-  },
-  "properties": {
-    "authenticationType": {
-      "kind": "property",
-      "displayName": "Authentication Type",
-      "group": "producer",
-      "label": "producer",
-      "required": false,
-      "type": "string",
-      "javaType": "java.lang.String",
-      "tags": [
-        "authentication-type"
-      ],
-      "deprecated": false,
-      "secret": false,
-      "componentProperty": true,
-      "defaultValue": "basic",
-      "enum": [
-        {
-          "value": "basic",
-          "label": "HTTP Basic Authentication"
-        }
-      ],
-      "description": "Type of authentication used to connect to the API"
-    },
-    "specification": {
-      "kind": "property",
-      "displayName": "Specification",
-      "group": "producer",
-      "label": "producer",
-      "required": false,
-      "type": "hidden",
-      "javaType": "java.lang.String",
-      "tags": [
-        "upload",
-        "url"
-      ],
-      "deprecated": false,
-      "secret": false,
-      "componentProperty": true,
-      "description": "Swagger specification of the service"
-    },
-    "basePath": {
-      "kind": "property",
-      "displayName": "Base path",
-      "group": "producer",
-      "label": "producer",
-      "required": false,
-      "type": "string",
-      "javaType": "java.lang.String",
-      "tags": [],
-      "deprecated": false,
-      "secret": false,
-      "componentProperty": true,
-      "defaultValue": "/api",
-      "description": "API basePath for example /v2. Default is unset if set overrides the value present in Swagger specification."
-    },
-    "host": {
-      "kind": "property",
-      "displayName": "Host",
-      "group": "producer",
-      "label": "producer",
-      "required": false,
-      "type": "string",
-      "javaType": "java.lang.String",
-      "tags": [],
-      "deprecated": false,
-      "secret": false,
-      "componentProperty": true,
-      "defaultValue": "http://sample.io",
-      "description": "Scheme hostname and port to direct the HTTP requests to in the form of https://hostname:port. Can be configured at the endpoint component or in the correspoding REST configuration in the Camel Context. If you give this component a name (e.g. petstore) that REST configuration is consulted first rest-swagger next and global configuration last. If set overrides any value found in the Swagger specification RestConfiguration. Can be overriden in endpoint configuration."
-    },
-    "username": {
-      "kind": "property",
-      "displayName": "Username",
-      "group": "producer",
-      "label": "producer",
-      "required": true,
-      "type": "string",
-      "javaType": "java.lang.String",
-      "deprecated": false,
-      "secret": false,
-      "componentProperty": true,
-      "description": "Username to authenticate with"
-    },
-    "password": {
-      "kind": "property",
-      "displayName": "Password",
-      "group": "producer",
-      "label": "producer",
-      "required": true,
-      "type": "string",
-      "javaType": "java.lang.String",
-      "deprecated": false,
-      "secret": true,
-      "componentProperty": true,
-      "description": "Password to authenticate with"
-    }
-  },
-  "actions": [
-    {
-      "id": "io.syndesis:rest-swagger-connector:@syndesis-connectors.version@:_id_:operation-0",
-      "name": "GET /operation",
-      "descriptor": {
-        "camelConnectorGAV": "io.syndesis:rest-swagger-connector:@syndesis-connectors.version@",
-        "camelConnectorPrefix": "swagger-operation",
-        "inputDataShape": {
-          "kind": "none"
-        },
-        "outputDataShape": {
-          "kind": "none"
-        },
-        "configuredProperties": {
-          "operationId": "operation-0"
-        }
-      },
-      "actionType": "connector",
-      "pattern": "To"
-    }
-  ]
-}
